Refactor MLKit camera functionality
