+++ app/src/main/java/openfoodfacts/github/scrachx/openfood/features/scan/WorkflowState.kt
@@ -1,0 +1,11 @@
+package openfoodfacts.github.scrachx.openfood.features.scan
+
+/**
+ * State set of the application workflow.
+ */
+enum class WorkflowState {
+    NOT_STARTED,
+    DETECTING,
+    DETECTED,
+    CONFIRMING
+}
