+++ app/src/playstore/java/openfoodfacts/github/scrachx/openfood/camera/WorkflowModel.kt
@@ -5,6 +5,7 @@ import androidx.annotation.MainThread
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.MutableLiveData
 import com.google.mlkit.vision.barcode.Barcode
+import openfoodfacts.github.scrachx.openfood.features.scan.WorkflowState
 
 /**
  * View model for handling application workflow based on camera preview.
@@ -16,16 +17,6 @@ class WorkflowModel(application: Application) : AndroidViewModel(application) {
 
     var isCameraLive = false
         private set
-
-    /**
-     * State set of the application workflow.
-     */
-    enum class WorkflowState {
-        NOT_STARTED,
-        DETECTING,
-        DETECTED,
-        CONFIRMING
-    }
 
     @MainThread
     fun setWorkflowState(workflowState: WorkflowState) {
